mailing_service/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.0.12
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.7.0
Pillow==9.2.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.26.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
celery
redis
